Fix for Order in extracted chunk undefined webpack errors
