--- src/shared/components/event-meta/style.css
@@ -1,4 +1,0 @@
-.eventLinks {
-  composes: cf from "../../components/utils/layout.css";
-  margin-bottom: 1em;
-}
