--- src/shared/components/event-title/style.css
@@ -1,27 +1,0 @@
-@value red from "../variables/colors.css";
-
-.eventTitleLink {
-  text-decoration: none;
-}
-
-.eventTitle {
-  font-size: 1.6em;
-  text-decoration: none;
-  display: inline;
-  margin: 0 0 20px 0;
-  float: left;
-  text-align: left;
-}
-
-.eventTitle span {
-  display: inline;
-}
-
-.eventTitle:hover, .eventTitle:active, .eventTitle:focus {
-  color: red;
-}
-
-.arrow {
-  color: red;
-  padding-left: 5px;
-}
