--- src/shared/components/events-list-entry/style.css
@@ -1,15 +1,0 @@
-.eventItem {
-  position: relative;
-  list-style-type: none;
-}
-
-.mobileHorizontalLine {
-  display: block;
-  margin: 1.5em 0 0 0;
-}
-
-.eventDescription {
-  margin-top: 10px;
-  float: left;
-}
-
