+++ src/shared/components/events-list/style.css
@@ -14,15 +14,60 @@
   text-align: center;
 }
 
+.eventsListTimelineSection h2.eventTitle {
+  text-align: left;
+}
+
 .eventsList {
   font-size: 1em;
   margin-top: 0;
   padding-left: 0;
 }
 
+.eventItem {
+  position: relative;
+  list-style-type: none;
+}
+
+.eventTitleLink {
+  text-decoration: none;
+}
+
+.eventTitle {
+  composes: h2 from "../../components/typography/style.css";
+  font-size: 1.6em;
+  text-decoration: none;
+  display: inline;
+  margin: 0 0 20px 0;
+  float: left;
+}
+
+.eventTitle span {
+  display: inline;
+}
+
+.eventTitle:hover, .eventTitle:active, .eventTitle:focus {
+  color: red;
+}
+
 .arrow {
   color: red;
   padding-left: 5px;
+}
+
+.eventDescription {
+  margin-top: 10px;
+  float: left;
+}
+
+.eventLinks {
+  composes: cf from "../../components/utils/layout.css";
+  margin-bottom: 1em;
+}
+
+.mobileHorizontalLine {
+  display: block;
+  margin: 1.5em 0 0 0;
 }
 
 @media mobile {
